+++ src/main/java/insane96mcp/progressivebosses/module/dragon/feature/ResistancesFeature.java
@@ -23,7 +23,7 @@ public class ResistancesFeature extends Feature {
 	private final ForgeConfigSpec.ConfigValue<Double> damageRedutionWhenSittingConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> explosionDamageReductionConfig;
 
-	public double damageRedutionWhenSitting = 0.035d;
+	public double damageRedutionWhenSitting = 0.08d;
 	public double explosionDamageReduction = 0.667d;
 
 	public ResistancesFeature(Module module) {