+++ changelog.md
@@ -11,7 +11,7 @@
     * When respawning crystals the Ender Dragon now takes a flat amount of bonus damage (~~+25%~+125% damage based off current health~~ -> +33% damage)
     * If the crystal to respawn is a decimal value (e.g. 0.6) the dragon now takes the decimal part as a chance to spawn a +1 crystal (so 60% chance to respawn 1 and 40% chance to fail the respawn)
   * Health
-    * Bonus health per difficulty reduced (~~103~~ -> 25 Health at max difficulty)
+    * Bonus health per difficulty reduced (~~103~~ -> 30 Health at max difficulty)
     * Removed bonus health regen from crystals (~~0.15 hp/s per difficulty~~ -> REMOVED)
   * Resistances and Vulnerabilities
     * Removed bonus damage when not in center (~~+1% current health as bonus indirect damage~~ -> REMOVED; ~~+3% current health as bonus direct damage~~ -> REMOVED)